% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbr.R
\docType{package}
\name{neighbr}
\alias{neighbr}
\alias{neighbr-package}
\title{neighbr: A package for computing nearest neighbors.}
\description{
Classification, regression, and clustering with k nearest neighbors.
}
\section{Neighbr functions}{

The package contains a function for running the nearest neighbors algorithm,
as well as functions to directly compute distances between two vectors.
}

\section{More information}{

\code{\link{knn}} documents the main \code{knn} function.

\code{\link{distance}} and \code{\link{similarity}} provide details on standalone measures.

For more details and examples, see the vignette by running the following:

\code{vignette("neighbr-help")}
}

